Canvas Rendered menu and graphics overhaul, replacement of old landing page
