--- alt_index.html
@@ -1,43 +1,0 @@
-
-<!-- const FOODCOLOUR = "#EF798A"; -->
-<!-- const TEXTCOLOUR = "#4F646F"; -->
-<!-- const PLAYER1COLOUR = "#86BBD8"; -->
-<!-- const PLAYER2COLOUR = "#BFFFBC"; -->
-<!-- const BGCOLOUR = "#120D31"; -->
-
-<!-- :root { -->
-<!--     --rose-pink: #EF798A; -->
-<!--     --dark-grey: #4F646F; -->
-<!--     --blue-grey: #86BBD8; -->
-<!--     --pastel-green: #BFFFBC; -->
-<!--     --dark-blue: #120D31; -->
-<!-- } -->
-
-<!DOCTYPE html>
-<html>
-<link rel="stylesheet" href="utils/shared.css">
-<style>
-    html, body {
-	background-color: var(--dark-blue);
-	margin: 0;
-	padding: 0;
-	overflow: hidden;
-	width: 100%;
-	height: 100%;
-    }
-
-    #canvas {
-	display: block;
-	width: 100vw;
-	height: 100vh;
-    }
-
-</style>
-
-<body>
-    <canvas id="canvas"> </canvas>
-    <script type="module" src="./alt_index.js"></script>
-    <script type="module" src="./utils/utils.js"></script>
-</body>
-
-</html>
