--- games/runner/index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <style>
-	html {
-	    background-color: black;
-	}
-
-	body {
-	    display: flex;
-	    flex-direction: column;
-	    justify-content: center;
-	    align-items: center;
-	    color: white;
-	}
-    </style>
-
-    <body>
-	<h1> RUNNER </h1>
-	<canvas id="canvas" height="600" width="800"> </canvas>
-	<script type="module" src="../../utils/utils.js"></script>
-	<script type="module" src="runner.js"></script>
-    </body>
-</html>
